:truck: 히스토리 테이블 features에서 widgets으로 이동
