+++ src/widgets/history-table/config.ts
