+++ src/widgets/history-table/ui/render-cell.tsx
